sweep fixes and invariant graph improvements
1. Fixed issues found when running the sweep
2. non-constrained leaf variables and their defining invariants are now recursively removed from the invariant graph
3. Improved testing by marking some methods that should only be used when doing local regression as deprecated, reducing the risk of reverting the local regression

+++ src/search/neighborhoods/countNeighborhood.cpp
@@ -54,8 +54,8 @@ size_t CountNeighborhood::randomMove(RandomProvider& random,
   if (_eligibleIndices.size() <= _amount) {
     return 0;
   }
-  _index1 = random.intInRange(0, _amount - 1);
-  _index2 = random.intInRange(_amount, _eligibleIndices.size() - 1);
+  _index1 = random.intInRange(0, static_cast<Int>(_amount) - 1);
+  _index2 = random.intInRange(static_cast<Int>(_amount), static_cast<Int>(_eligibleIndices.size()) - 1);
   _curTimestamp = assignment.currentTimestamp();
   assignment.set(
       _vars[_eligibleIndices[_index1]].solverId(),
